style(input): fix the header
The license notice must not change. The only information to be updated is the date of modification, the author and the copyright notice.

Why license notices?
A program must have, in addition to a copyright notice, an explicit statement that “This program is released under license FOO,”
that leaves no room for uncertainty about whether the license FOO applies to the code of that program,
even if the program has a copy of a license FOO alongside the source files.