refactor: improve special keys handling logic
Completely redone the logic for handling special keys. Built an array of structs containing a pair of key_value-function_pointer. The array is sorted and accessed in O(log n). If the key is missing, return 0. If the key handler has been accessed, return 1.

Previously, the logic was based on if/else and switch/case statements, which added too much complexity.